+++ ChangeLog.md
@@ -6,6 +6,7 @@
 
 - Added `GeminiURI.without_query`.
   ([#32](https://github.com/davep/wasat/pull/32))
+- Added `GeminiURI.parent` ([#32](https://github.com/davep/wasat/pull/32))
 
 ## v0.7.0